feat(sitemap): added a default-sitemap.xml & plugins-sitemap.xml (fetched from API)
closes #741
