+++ components/layout/Container.vue
@@ -30,148 +30,14 @@
     import Breadcrumb from "~/components/layout/Breadcrumb.vue";
     import NavToc from "~/components/docs/NavToc.vue";
     import {hash} from "ohash";
-
-    const config = useRuntimeConfig();
-
-    const generateSubMenu = (baseUrl, group, items) => generateSubMenuWithGroupProvider(baseUrl, () => group, items);
-
-    const generateSubMenuWithGroupProvider = (baseUrl, groupProviderFromItem, items) => {
-        let itemsBySubmenu = items.reduce((m, item) => {
-            const subMenuSplitter = item.lastIndexOf(".");
-            if (subMenuSplitter === -1) {
-                m[item] = {
-                    title: toNavTitle(item),
-                    _path: `${baseUrl}/${groupProviderFromItem(item)}.${item}`
-                }
-            } else {
-                let subMenuName = item.substring(0, subMenuSplitter);
-                if (!m[subMenuName]) {
-                    const subGroup = `${groupProviderFromItem(item)}.${subMenuName}`;
-                    m[subMenuName] = generateSubMenu(
-                        `${baseUrl}/${subMenuName}`,
-                        subGroup,
-                        items.filter(i => i.startsWith(subMenuName + "."))
-                            .map(i => i.substring(subMenuName.length + 1))
-                    );
-                }
-            }
-            return m;
-        }, {});
-
-        return Object.entries(itemsBySubmenu).map(([key, value]) => {
-            if (Array.isArray(value)) {
-                return {
-                    title: toNavTitle(key),
-                    _path: `${baseUrl}/${key}`,
-                    isPage: false,
-                    children: value
-                }
-            }
-            return value;
-        });
-    }
-
-    const toNavTitle = (title) => title.split(".")
-        .map(string => string.charAt(0).toUpperCase() + string.slice(1))
-        .join("");
-
-    const recursivePages = (item) => {
-        const paths = [];
-        if (item.isPage ?? true) {
-            paths.push(item._path);
-        }
-        if (item.children) {
-            paths.push(...(item.children.flatMap(child => {
-                return recursivePages(child);
-            })));
-        }
-
-        return paths;
-    }
+    import {recursivePages} from "~/utils/navigation.js";
 
     const fetchNavigation = async () => {
         const queryBuilder = queryContent('/' + props.type + '/').without("body");
 
         let navigationFetch;
         if (props.type === "plugins") {
-            const categories = ['tasks', 'triggers', 'conditions', 'controllers', 'storages', 'secrets', 'guides'];
-            navigationFetch = await useAsyncData(
-                'plugins-nav',
-                () => $fetch(`${config.public.apiUrl}/plugins`),
-                {
-                    transform: plugins => {
-                        let sortedPluginsHierarchy = plugins.map(plugin => {
-                            let rootPluginUrl = '/plugins/' + plugin.name;
-                            const children = categories
-                                .flatMap(category => {
-                                    let children;
-                                    if (plugin.name === "core") {
-                                        if (category === "tasks") {
-                                            children = generateSubMenu(
-                                                `${rootPluginUrl}/${category}`,
-                                                plugin.group,
-                                                plugin[category].map(item => item.split(".").slice(-2).join("."))
-                                            );
-                                        } else {
-                                            const fqnByClassName = {};
-                                            children = generateSubMenuWithGroupProvider(
-                                                `${rootPluginUrl}/${category}`,
-                                                (item) => {
-                                                    const fqn = fqnByClassName[item];
-
-                                                    return fqn.substring(0, fqn.lastIndexOf("."));
-                                                },
-                                                plugin[category].map(item => {
-                                                    const className = item.substring(item.lastIndexOf(".") + 1);
-                                                    fqnByClassName[className] = item;
-                                                    return className;
-                                                })
-                                            );
-                                        }
-                                    } else {
-                                        children = generateSubMenu(
-                                            `${rootPluginUrl}/${category}`,
-                                            `${plugin.group}`,
-                                            plugin[category].map(item => item.substring(plugin.group.length + 1))
-                                        );
-                                    }
-
-                                    if (children.length === 0) {
-                                        return [];
-                                    }
-
-                                    return {
-                                        title: toNavTitle(category),
-                                        _path: `${rootPluginUrl}/${category}`,
-                                        isPage: false,
-                                        children
-                                    }
-                                });
-                            return {
-                                title: toNavTitle(plugin.title),
-                                _path: rootPluginUrl,
-                                children
-                            };
-                        }).sort((a, b) => {
-                            const nameA = a.title.toLowerCase(),
-                                nameB = b.title.toLowerCase();
-
-                            if (nameA === "core") {
-                                return -1;
-                            }
-                            if (nameB === "core") {
-                                return 1;
-                            }
-
-                            return nameA === nameB ? 0 : nameA < nameB ? -1 : 1;
-                        });
-                        return [{
-                            title: "Plugins",
-                            _path: "/plugins",
-                            children: sortedPluginsHierarchy
-                        }];
-                    }
-                });
+            navigationFetch = await useFetch(`/api/plugins?type=navigation`);
         } else {
             navigationFetch = await useAsyncData(
                 `NavSideBar-${hash(props.type)}`,
